+++ crates/kelk/README.md
@@ -1,0 +1,3 @@
+# Kelk
+
+Kelk provide basic materials for writhing WASM based contracts in [Zarb](https://zarb.network/) blockchain.
