+++ crates/kelk/src/blockchain/mod.rs
@@ -2,5 +2,6 @@
 
 
 pub mod blockchain;
+pub mod mock;
 
 pub use blockchain::Blockchain;
