+++ crates/kelk/src/storage/collections/vector/error.rs
