+++ crates/kelk/src/storage/mock.rs
