+++ examples/calculator/Cargo.toml
@@ -5,8 +5,7 @@ edition = "2021"
 authors = ["Zarb development team <admin@zarb.network>"]
 
 [dependencies]
-kelk-env = { path = "../../kelk/env" }
-kelk-lib = { path = "../../kelk/lib" }
+kelk = { path = "../../crates/kelk" }
 minicbor = { version = "0.18.0", features = ["half", "derive"] }
 wee_alloc = { version = "0.4", default-features = false }
