--- kelk/lib/README.md
@@ -1,3 +1,0 @@
-# Kelk-lib
-
-Kelk-lib is the standard library for wasm based contracts in [Zarb](https://zarb.network/) blockchain.
